first changes v5

+++ Aiva.Core/Database/Users.cs
@@ -15,23 +15,21 @@ namespace Aiva.Core.Database {
             /// Add or Update User in Database
             /// </summary>
             /// <param name="User">todo: describe User parameter on AddUser</param>
-            private static void AddUserToDatabase(TwitchLib.Models.API.User.User User) {
+            private static void AddUserToDatabase(TwitchLib.Models.API.v5.Users.User User) {
 
                 if (User != null) {
                     using (var context = new Storage.StorageEntities()) {
-                        var databaseUser = context.Users.SingleOrDefault(u => u.Id == User.Id);
+                        var databaseUser = context.Users.SingleOrDefault(u => String.Compare(u.Id, User.Id) == 0);
 
                         // Update User
                         if (databaseUser != null) {
                             databaseUser.Bio = User.Bio;
-                            databaseUser.CreatedAt = User.CreatedAt.ToString();
+                            databaseUser.CreatedAt = User.CreatedAt;
                             databaseUser.DisplayName = User.DisplayName;
                             databaseUser.Logo = User.Logo;
                             databaseUser.Name = User.Name;
-                            databaseUser.TimeSinceCreated = User.TimeSinceCreated.ToString();
-                            databaseUser.TimeSinceUpdated = User.TimeSinceUpdated.ToString();
                             databaseUser.Type = User.Type;
-                            databaseUser.UpdatedAt = User.UpdatedAt.ToString();
+                            databaseUser.UpdatedAt = User.UpdatedAt;
 
                             // Active users
                             if (databaseUser.ActiveUsers == null) {
@@ -45,16 +43,14 @@ namespace Aiva.Core.Database {
                         // Create User
                         else {
                             var newUser = new Storage.Users {
-                                Id = User.Id.Value,
+                                Id = User.Id,
                                 Bio = User.Bio,
-                                CreatedAt = User.CreatedAt.ToString(),
+                                CreatedAt = User.CreatedAt,
                                 DisplayName = User.DisplayName,
                                 Logo = User.Logo,
                                 Name = User.Name,
-                                TimeSinceCreated = User.TimeSinceCreated.ToString(),
-                                TimeSinceUpdated = User.TimeSinceUpdated.ToString(),
                                 Type = User.Type,
-                                UpdatedAt = User.UpdatedAt.ToString(),
+                                UpdatedAt = User.UpdatedAt,
                                 Currency = new Storage.Currency {
                                     Value = 0,
                                 },
@@ -97,21 +93,27 @@ namespace Aiva.Core.Database {
             /// </summary>
             /// <param name="sender">todo: describe sender parameter on RemoveUserFromActiveUsers</param>
             /// <param name="e">todo: describe e parameter on RemoveUserFromActiveUsers</param>
-            public static void RemoveUserFromActiveUsers(object sender, OnUserLeftArgs e) {
+            public async static void RemoveUserFromActiveUsers(object sender, OnUserLeftArgs e) {
                 using (var context = new Storage.StorageEntities()) {
-                    var twitchID = TwitchApi.Users.GetUser(e.Username);
+                    //var twitchID = TwitchApi.Users.GetUser(e.Username);
+                    var twitchID = await TwitchAPI.Users.v5.GetUserByName(e.Username);
 
-                    if (twitchID != null) {
-                        var user = context.Users.SingleOrDefault(u => u.Id == twitchID.Id.Value);
+                    if (twitchID != null && twitchID.Total > 0) {
+                        foreach (var userMatch in twitchID.Matches) {
+                            if (String.Compare(userMatch.Name, e.Username, true) != 0)
+                                continue;
 
-                        if (user != null) {
-                            var duration = DateTime.Now.Subtract(user.ActiveUsers.JoinedTime.Value);
+                            var user = context.Users.SingleOrDefault(u => String.Compare(u.Id, userMatch.Id) == 0);
 
-                            user.TimeWatched.TimeWatched1 += duration.Ticks;
+                            if (user != null) {
+                                var duration = DateTime.Now.Subtract(user.ActiveUsers.JoinedTime.Value);
 
-                            context.ActiveUsers.Remove(user.ActiveUsers);
+                                user.TimeWatched.TimeWatched1 += duration.Ticks;
 
-                            context.SaveChanges();
+                                context.ActiveUsers.Remove(user.ActiveUsers);
+
+                                context.SaveChanges();
+                            }
                         }
                     }
                 }
@@ -124,7 +126,7 @@ namespace Aiva.Core.Database {
         /// <param name="username"></param>
         /// <returns></returns>
         public static bool IsUserInDatabase(string username) {
-            using(var context = new Storage.StorageEntities()) {
+            using (var context = new Storage.StorageEntities()) {
                 return context.Users.Any(u => String.Compare(u.Name, username, true) == 0);
             }
         }
